Cubes display

+++ src/buffer/VertexBuffer.cpp
@@ -51,7 +51,7 @@ void VertexBuffer::init()
 	// Start of user code init
     Buffer::init();
     glBindBuffer(GL_ARRAY_BUFFER, id);
-    glBufferData(GL_ARRAY_BUFFER, sizeof(GLfloat)*data->size(), &(*data)[0], GL_STATIC_DRAW);
+    glBufferData(GL_ARRAY_BUFFER, sizeof(unsigned int)*data->size(), &(*data)[0], GL_STATIC_DRAW);
     glBindBuffer(GL_ARRAY_BUFFER, 0);
 	// End of user code
 }
